WIP: Dynamically create test data
Reimplement AssemblyDocumentationTest using dynamically generated assembly.
