--- src/MdDocs.ApiReference.Model.Test/AssemblyDocumenationTest.cs
@@ -1,39 +1,0 @@
-﻿using Grynwald.MdDocs.ApiReference.Model;
-using Grynwald.MdDocs.ApiReference.Test.TestData;
-using Xunit;
-
-namespace Grynwald.MdDocs.ApiReference.Test.Model
-{
-    public class AssemblyDocumenationTest : TestBase
-    {
-        [Fact]
-        public void TryGetDocumentation_returns_null_for_an_undocumented_type()
-        {
-            // ARRANGE
-            var typeId = GetTypeId(typeof(TestClass_InternalType));
-            var sut = m_AssemblyDocumentation.Value;
-
-            // ACT
-            var documentation = sut.TryGetDocumentation(typeId);
-
-            // ASSERT
-            Assert.Null(documentation);
-        }
-
-        [Fact]
-        public void TryGetDocumentation_returns_expected_documentation_item_for_an_documented_type()
-        {
-            // ARRANGE
-            var typeId = GetTypeId(typeof(TestClass_Type));
-            var sut = m_AssemblyDocumentation.Value;
-
-            // ACT
-            var documentation = sut.TryGetDocumentation(typeId);
-
-            // ASSERT
-            Assert.NotNull(documentation);
-            var typeDocumentation = Assert.IsType<TypeDocumentation>(documentation);
-            Assert.Equal(typeId, typeDocumentation.TypeId);
-        }
-    }
-}
